+++ tools/webhook.js
@@ -3,6 +3,8 @@ const crypto = require('crypto');
 const child_process = require('child_process');
 const fs = require('fs');
 
+const MANUAL_DEPLOY_URL = '?do=deploy';
+
 // ----------------------------------------------------------------------------
 //   Functions/Methods/etc:
 // ----------------------------------------------------------------------------
@@ -13,6 +15,7 @@ const fs = require('fs');
  */
 function deploy(target) {
   let command = `cd ${process.env.GLOBAL_121_REPO} && sudo ./tools/deploy.sh`;
+  target = sanitizeTarget(target);
 
   if (target) {
     command += ` "${target}"`;
@@ -49,35 +52,79 @@ function isMinorUpgrade(target) {
   return target.includes(currentMinorVersion);
 }
 
+function showManualDeployForm() {
+  return `<!DOCTYPE html>
+<meta charset="utf-8">
+<title></title>
+<form method="POST" action="${MANUAL_DEPLOY_URL}">
+<input name="secret" type="password">
+<input name="target">
+<input type="submit">
+</form>`;
+}
+
+function sanitizeTarget(input) {
+  return input.replace(/[^a-z/-_.0-9]/gi, '');
+}
+
 // ----------------------------------------------------------------------------
 //   Webhook Service:
 // ----------------------------------------------------------------------------
 
 http
-  .createServer(function (req, res) {
+  .createServer((req, res) => {
+    if (req.method === 'GET' && req.url.endsWith(MANUAL_DEPLOY_URL)) {
+      console.log(`Show manual interface.`);
+      res.end(showManualDeployForm());
+      return;
+    }
+
     let body = [];
     req.on('data', function (chunk) {
       body.push(chunk);
     });
     req.on('end', function () {
-      let str = Buffer.concat(body).toString();
-      let sig =
-        'sha1=' +
-        crypto
-          .createHmac('sha1', process.env.GITHUB_WEBHOOK_SECRET)
-          .update(str)
-          .digest('hex');
-      let payload = JSON.parse(str);
+      let strBody = Buffer.concat(body).toString();
+      let payload = {};
 
-      if (req.headers['x-hub-signature'] !== sig) {
-        console.warn('Invalid GitHub signature!');
+      if (req.method === 'POST' && req.url.endsWith(MANUAL_DEPLOY_URL)) {
+        console.log(`Process manual request`);
+        payload = new URL('http://example.org/?' + strBody).searchParams;
+
+        if (payload.get('secret') !== process.env.DEPLOY_SECRET) {
+          console.warn('Secret does not match.');
+          return;
+        }
+        let target = sanitizeTarget(payload.get('target'));
+        if (!target) {
+          console.log(`No valid target.`);
+          return;
+        }
+        console.log(`Manual deployment for: ${target} `);
+        deploy(target);
         return;
+      }
+
+      if (req.headers['x-hub-signature']) {
+        const sig =
+          'sha1=' +
+          crypto
+            .createHmac('sha1', process.env.GITHUB_WEBHOOK_SECRET)
+            .update(strBody)
+            .digest('hex');
+
+        if (req.headers['x-hub-signature'] !== sig) {
+          console.warn('Invalid GitHub signature!');
+          return;
+        }
+
+        payload = JSON.parse(strBody);
       }
 
       if (
         payload.pull_request &&
         payload.pull_request.merged &&
-        payload.pull_request.title.includes('[SKIP CD]')
+        payload.pull_request.title.toUpperCase().includes('[SKIP CD]')
       ) {
         console.log('PR deployment skipped with [SKIP CD]');
         return;
@@ -94,9 +141,9 @@ http
       }
 
       if (
-        process.env.DEPLOY_PRE_RELEASE &&
-        payload.action === 'prereleased' &&
+        !!process.env.DEPLOY_PRE_RELEASE &&
         process.env.NODE_ENV === 'production' &&
+        payload.action === 'prereleased' &&
         payload.release.draft === false &&
         payload.release.prerelease === true &&
         payload.release.target_commitish
@@ -109,12 +156,14 @@ http
       }
 
       if (
-        process.env.DEPLOY_RELEASE &&
-        payload.action === 'released' &&
+        !!process.env.DEPLOY_RELEASE &&
         process.env.NODE_ENV === 'production' &&
+        payload.action === 'released' &&
         payload.release.draft === false &&
         payload.release.target_commitish &&
-        isMinorUpgrade(payload.release.target_commitish)
+        (!!process.env.DEPLOY_PATCH
+          ? isMinorUpgrade(payload.release.target_commitish)
+          : true)
       ) {
         console.log(
           `Release (hotfix) deployment for: ${payload.release.target_commitish}`,